Update image to boot from usb-drive and UEFI
By adding some bootloader code during the image generation, we are now able
to write the image to usb-drive (via dd i.e.) and use it to boot.
Also we are now able to boot with UEFI. To get the installation to work, we
also need to download the signed kernel packages.